Remove usage of LightCouch.
It is not working properly with sonews' Article serialization.
A thin layer (based on Apache HttpComponents) between sonews and CouchDB
will be a better choice. jcouchdb is such a wrapper (although not that
lightweight) but it is unmaintained for years.

I've added a CouchDBClient stub which should later provide basic CouchDB
API operations such as PUT a document or call a VIEW.